look at all warnings
