+++ src/NtFreX.Audio.Alsa/AlsaCaptureContext.cs
@@ -1,13 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace NtFreX.Audio.Alsa
+﻿namespace NtFreX.Audio.Alsa
 {
     /// <summary>
     /// https://github.com/ZhangGaoxing/alsa.net
     /// </summary>
-    class AlsaCaptureContext
+    public class AlsaCaptureContext
     {
     }
 }
