+++ src/NtFreX.Audio.Alsa/Interop/Interop.alsa.cs
@@ -12,7 +12,6 @@ namespace NtFreX.Audio.Alsa.Interop
         [DllImport(AlsaLibrary)]
         internal static extern IntPtr snd_strerror(int errnum);
 
-        /// <param name="pcm">snd_pcm_t**</param>
         [DllImport(AlsaLibrary)]
         internal static extern HResult snd_pcm_open(ref IntPtr pcm, string name, snd_pcm_stream_t stream, int mode);
 
@@ -46,7 +45,6 @@ namespace NtFreX.Audio.Alsa.Interop
         [DllImport(AlsaLibrary)]
         internal static extern int snd_pcm_set_params(IntPtr pcm, snd_pcm_format_t format, snd_pcm_access_t access, uint channels, uint rate, int soft_resample, uint latency);
 
-        /// <param name="params">snd_pcm_hw_params_t**</param>
         [DllImport(AlsaLibrary)]
         internal static extern HResult snd_pcm_hw_params_malloc(out IntPtr @params);
 
@@ -59,7 +57,6 @@ namespace NtFreX.Audio.Alsa.Interop
         [DllImport(AlsaLibrary)]
         internal static extern HResult snd_pcm_hw_params_set_format(IntPtr pcm, IntPtr @params, snd_pcm_format_t val);
 
-        /// <param name="val">1 for mono and 2 for stereo</param>
         [DllImport(AlsaLibrary)]
         internal static extern HResult snd_pcm_hw_params_set_channels(IntPtr pcm, IntPtr @params, uint val);
 
@@ -69,14 +66,12 @@ namespace NtFreX.Audio.Alsa.Interop
         [DllImport(AlsaLibrary)]
         internal static extern HResult snd_pcm_hw_params(IntPtr pcm, IntPtr @params);
 
-        /// <param name="frames">snd_pcm_uframes_t*</param>
         [DllImport(AlsaLibrary)]
         internal static extern int snd_pcm_hw_params_get_period_size(IntPtr @params, out ulong frames, out IntPtr dir);
 
         [DllImport(AlsaLibrary)]
         internal static extern int snd_pcm_hw_params_set_period_size_near(IntPtr pcm, IntPtr @params, ref ulong frames, out IntPtr dir);
 
-        /// <param name="mixer">snd_mixer_t**</param>
         [DllImport(AlsaLibrary)]
         internal static extern int snd_mixer_open(ref IntPtr mixer, int mode);
 
@@ -92,21 +87,18 @@ namespace NtFreX.Audio.Alsa.Interop
         [DllImport(AlsaLibrary)]
         internal static extern int snd_mixer_selem_register(IntPtr mixer, IntPtr options, IntPtr classp);
 
-        /// <returns>snd_mixer_elem_t*</returns>
         [DllImport(AlsaLibrary)]
         internal static extern IntPtr snd_mixer_first_elem(IntPtr mixer);
 
         [DllImport(AlsaLibrary)]
         internal static extern IntPtr snd_mixer_elem_next(IntPtr elem);
 
-        /// <param name="elem">snd_mixer_elem_t*</param>
         [DllImport(AlsaLibrary)]
         internal static extern string snd_mixer_selem_get_name(IntPtr elem);
 
         [DllImport(AlsaLibrary)]
         internal static extern void snd_mixer_selem_id_alloca(IntPtr ptr);
 
-        /// <param name="mixer">snd_mixer_elem_t*</param>
         [DllImport(AlsaLibrary)]
         internal static extern int snd_mixer_selem_get_playback_volume(IntPtr elem, snd_mixer_selem_channel_id channel, out long value);
